requirements.txt
fastapi
uvicorn
langchain-core
langchain-openai
langgraph
python-dotenv
chromadb
